+++ src/test/java/backend/sumnail/medium/NailShopControllerTest.java
@@ -1,4 +1,4 @@
-package backend.sumnail.domain.nail_shop.controller;
+package backend.sumnail.medium;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
